requirements.txt
PyYAML~=5.3.1
pandas~=1.3.5
mariadb~=1.0.11
sqlalchemy~=1.4.35
urllib3~=1.26.8
requests~=2.27.1
plotly~=5.7.0
